+++ Assets/Editor Toolbox/Editor/Utilities/GuiLayoutUtility.cs
@@ -13,7 +13,6 @@ namespace Toolbox.Editor
         /// </summary>
         public static readonly float layoutPadding = -2 * EditorGUIUtility.standardVerticalSpacing;
 
-
         public static void BeginStrechedVertical()
         {
             BeginFixedVertical(GUIStyle.none);
@@ -54,6 +53,26 @@ namespace Toolbox.Editor
             GUILayout.Space(space);
         }
 
-        //TODO: add more helper methods
+        /// <summary>
+        /// A bit hacky way to retrieve the width of currently active layout.
+        /// Width is properly calculated only when the <see cref="Event.type"/> equals to <see cref="EventType.Repaint"/>.
+        /// </summary>
+        public static bool TryGetLayoutWidth(out float width)
+        {
+            using (var scope = new EditorGUILayout.HorizontalScope())
+            {
+                if (Event.current.type == EventType.Repaint)
+                {
+                    var scopeRect = scope.rect;
+                    width = scopeRect.width;
+                    return true;
+                }
+                else
+                {
+                    width = 0.0f;
+                    return false;
+                }
+            }
+        }
     }
 }