+++ Assets/Examples/Scripts/SampleBehaviour4.cs
@@ -72,15 +72,16 @@ public class SampleBehaviour4 : MonoBehaviour
 
     [Label("Horizontal Layout", skinStyle: SkinStyle.Box)]
 
-    [BeginHorizontal(labelToWidthRatio: 0.1f)]
+    [BeginHorizontal(LabelWidth = 50.0f)]
     public int var29;
+    [SpaceArea(10)]
     public int var30;
     [EndHorizontal]
     public int var31;
 
     [Label("Horizontal Layout (Group)", skinStyle: SkinStyle.Box)]
 
-    [BeginHorizontalGroup(label: "Horizontal Group")]
+    [BeginHorizontalGroup(Label = "Horizontal Group", ControlFieldWidth = true, ElementsInLayout = 2)]
     [ReorderableList(Foldable = true), InLineEditor]
     public GameObject[] gameObjects;
     [SpaceArea]